cJSON: 同步上游 clang 编译修复（LLONG_MAX 显式转 double）

+++ cJSON/cJSON.c
@@ -413,7 +413,7 @@ loop_end:
         }
     }
     /* use saturation in case of overflow */
-    else if (number >= LLONG_MAX)
+    else if (number >= (double)LLONG_MAX)
     {
         item->valueint = LLONG_MAX;
     }
@@ -440,7 +440,7 @@ CJSON_PUBLIC(double) cJSON_SetNumberHelper(cJSON *object, double number)
         return (double)NAN;
     }
 
-    if (number >= LLONG_MAX)
+    if (number >= (double)LLONG_MAX)
     {
         object->valueint = LLONG_MAX;
     }
@@ -2568,7 +2568,7 @@ CJSON_PUBLIC(cJSON *) cJSON_CreateNumber(double num)
         item->valuedouble = num;
 
         /* use saturation in case of overflow */
-        if (num >= LLONG_MAX)
+        if (num >= (double)LLONG_MAX)
         {
             item->valueint = LLONG_MAX;
         }
